PGP: Reassign things owned by user before dropping that user (#6)
Postgres will refuse to drop users who still own tables.  We will need
to reassign ownership before dropping it.  However, the Postgres
REASSIGN OWNED BY statement only affects things in the current database;
in order for that to work correctly, we will need to reconnect to
postgres with the correct database in the parameters.